Added NSFW warning to gifs from NSFW posts and subs
Now marks potentially NSFW gifs as such which should be helpful to those
who are going through the bot's comments and may not want to see such
content.

This includes a reworking of the context object, which is now given to
the gif_host object. This should help with context oriented decisions
further in the pipeline. There are also some small updates to the imgur
setup procedure to work with the new credentials object.