Massive overhaul of Gif Host library
Gif Host library is now much more modular and dynamic. Upload decision-
making is now done by the GifHostManager rather than through
predetermined switch statements in each host, centralizing complexity.
GifHosts simply define several parameters and GifHostManager decides
what to do with it.

Overall, this should centralize the decision-making of the
system, simply the implementation of each host, increase the usability
of the library, and generally be a better abstraction.

The completion of this overhaul will bring us to v3.0 and will
(hopefully) be the last major restructure of the bot.