Rewrote media metadata grabbing tools
And a todo message for context

MediaInfo is a new class that grabs all of the necessary metadata in
one or a few calls to ffprobe rather than several for each piece of
metadata. It streamlines the whole approach to grabbing metadata,
making it more consistent and faster than the old approach. This was
originally written in the vredditshare branch but there are serious
merging issues happening right now so right now I'm just porting things
to GifReversingBot and working from there.