Move host regex to their own classes
Slowly etching away at the remaining strongly coupled parts remaining from before the new gif host system. The gif host system is inherently weakly coupled so the last vestiges of the strong coupling should be removed for the sake of consistency at least. At best, moving the gif host system into it's own project might make it easier to update for GRB and VRS and allow for others to use it too.